refactor(astro-kbve): hoist GitHub wizard to /agents/github + shared guild picker
Reorganizes the agents surface from bot-centric to provider-centric.
GitHub is not a discordsh-specific concern — it's a cross-cutting
provider whose Vault rows (github_pat:<guild> + github_webhook:<guild>
+ github_repos:<guild>) are consumed by discordsh today and by any
future bot tomorrow. Nesting it under /agents/discordsh/github/
implied otherwise.

Route changes

- New: src/content/docs/dashboard/agents/github/index.mdx
- Removed: src/content/docs/dashboard/agents/discordsh/github/index.mdx
- Sidebar block under Agents becomes:
    Overview · GitHub · discordsh

Component refactor

- New: src/components/dashboard/ReactAgentGuildPicker.tsx — the guild
  picker is now its own component, consumed by both the discordsh
  manage page and the GitHub wizard. Reads / writes the same
  agentsService nanostores so picking a guild on one page persists
  to the other.
- ReactAgentDiscordsh.tsx drops its inline GuildPicker + the helpers
  it owned (guildIconUrl, initials, DISCORD_CDN — all moved into the
  shared component) and imports the new ReactAgentGuildPicker.
- ReactAgentGithubWizard.tsx replaces the previous "pick a guild on
  the discordsh page first" empty state with the same shared picker;
  the wizard now also surfaces the picker above the four steps so
  users can re-pick without bouncing between pages.

Catalog (P0)

- AstroAgentsDashboard.astro: new CardKind ('bot' | 'provider'),
  new GitHub provider card alongside the existing discordsh bot
  card. Each card surfaces both a kind badge (Provider / Bot) and
  the existing status badge (Available / Coming soon).
- discordsh card description tightened — points at the new GitHub
  provider page for its credential source.

Cross-link

- AstroAgentDiscordsh.astro: cross-link paragraph now reads
  "Run the GitHub provider wizard" → /dashboard/agents/github/
  (was /dashboard/agents/discordsh/github/).

No version bump

- axum-kbve v1.0.171 from the P0 PR has not yet published. P1
  (#11368), P2 (#11372), P3 (#11385) and now this refactor all
  ride the same image tag.

Out of scope (continues #11362)

- P4 per-guild repo allowlist (env GH_WEBHOOK_ALLOWED_REPOS →
  Vault github_repos:<guild>) + UI editor under the new GitHub
  provider page.
- P5 queue stats widget under /agents/github/.
- P6 bot install wizard.
- Future: hoist the token list itself off /agents/discordsh/ to a
  shared /agents/tokens/ surface so deleting a row provisioned via
  the GitHub provider page doesn't feel like it lives in the wrong
  spot.

Verification

- ./kbve.sh -nx astro-kbve:build ✓
- No new dependencies
