+++ apps/kbve/astro-kbve/astro.config.mjs
@@ -193,8 +193,8 @@ export default defineConfig({
 							collapsed: true,
 							items: [
 								{ label: 'Overview', link: '/dashboard/agents/', attrs: { 'data-auth-visibility': 'auth' } },
+								{ label: 'GitHub', link: '/dashboard/agents/github/', attrs: { 'data-auth-visibility': 'auth' } },
 								{ label: 'discordsh', link: '/dashboard/agents/discordsh/', attrs: { 'data-auth-visibility': 'auth' } },
-								{ label: 'discordsh · GitHub', link: '/dashboard/agents/discordsh/github/', attrs: { 'data-auth-visibility': 'auth' } },
 							],
 						},
 						{ label: 'Marketplace', link: '/dashboard/market/', attrs: { 'data-auth-visibility': 'auth' } },
