+++ apps/kbve/astro-kbve/src/content/docs/dashboard/agents/github/index.mdx
@@ -1,19 +1,19 @@
 ---
-title: discordsh · GitHub setup
-description: Guided wizard — generate HMAC webhook secret, register GitHub PAT, configure webhook, smoke-test backfill.
+title: GitHub provider
+description: Per-guild GitHub integration — HMAC webhook secret, PAT, webhook configuration, smoke-test backfill. Shared by every KBVE agent that consumes the GitHub Vault rows.
 tableOfContents: false
 graph:
     visible: false
 sidebar:
-    label: GitHub setup
+    label: GitHub
     order: 2
 unsplash: 1558494949-ef010cbdcc31
 img: https://images.unsplash.com/photo-1558494949-ef010cbdcc31?fit=crop&w=1400&h=700&q=75
 tags:
     - dashboard
     - agents
-    - discordsh
     - github
+    - integrations
 ---
 
 import AstroAgentGithubWizard from '@/components/dashboard/AstroAgentGithubWizard.astro';
